etl_twitter_airflow/requirements.txt
aiobotocore==2.4.0
aiohttp==3.8.3
aioitertools==0.11.0
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
botocore==1.27.59
certifi==2022.9.24
charset-normalizer==2.1.1
frozenlist==1.3.1
fsspec==2022.8.2
idna==3.4
jmespath==1.0.1
multidict==6.0.2
numpy==1.23.3
oauthlib==3.2.1
pandas==1.5.0
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.4
requests==2.28.1
requests-oauthlib==1.3.1
s3fs==2022.8.2
six==1.16.0
tweepy==4.10.1
typing-extensions==4.3.0
urllib3==1.26.12
wrapt==1.14.1
yarl==1.8.1
